Fix 048
Fix 048:Show/Hide option for empty annotations under OPTION need to get
reworked.   We can work on speed of this later if initially an issue
(because you have to trash and re-render comment pane for that
paragraph).
Mar 22: Fixed. Use css approach so speed won’t be an issue. (Zhenwei)