Fix more bugs: 0058-0064 (on Annotation program software changes)
Fixed more bugs found in testing. Which mainly includes the following
two:

0060: 04/30/2016, Shenmeng
http://secure-dusk-1121.herokuapp.com/version8b/a3?user=andy&db=annotati
onplaypen
We cannot see the annotations we make immediately when we make
annotations not on top of other annotations. New annotations are not ave
We cannot see them even if we click on them afterwards. We could only
refresh to see them.
Zhenwei:  Fixed. Now new annotation buttons should pop out as expected.
Also disabled the click function under select mode. The user can only
review the annotation by clicking the button on the right.

0059: 04/30/2016, Shenmeng
http://secure-dusk-1121.herokuapp.com/version8b/a3?user=andy&db=annotati
onplaypen
When clicking on some single-line paragraphs, the annotation panel
doesn’t display. Only when we click on the annotation icon the right
side of the text could the annotation panel  be displayed.
For instance, while “Literature review” works, the line below, the
“Academic Users' Information-Seeking and Searching Behaviors” doesn’t.
Other paragraphs like this include: “Information Sources for Scholarly
Information”, “RQ1. What Are the Characteristics of a Research Task?”,
“Explorative: The topic of a research task was relatively new in the
field.”, “Uncertain: The topic of a research task lacked certainty.”,
etc.